Implement Cobra Command and Viper options handling
ENV format is `R8GET_<OPTION_NAME>` where option-name is translated to
all caps *and* `-` is replaced with `_`. E.g. `minimum-chunk-size` becomes
`R8GET_MINIMUM_CHUNK_SIZE`.

This commit restructures the code somewhat to isolate concerns and adds
the following packages:

* confg - config specific handling

* optname - constants for string option names
            allows for consistent naming when referencing via Viper

* download - download-specific code

* extract - tar extraction-specific code

This commit lays the groundwork for up coming larger changes.
