+++ optname/names.go
@@ -1,0 +1,10 @@
+package optname
+
+const (
+	Concurrency      = "concurrency"
+	Extract          = "extract"
+	Force            = "force"
+	MinimumChunkSize = "minimum-chunk-size"
+	Retries          = "retries"
+	Verbose          = "verbose"
+)
